Split form files
