+++ packages/base/src/index.ts
@@ -4,7 +4,7 @@ export * from './mainview';
 export * from './tools';
 export * from './types';
 export * from './widget';
-export * from './formbuilder';
+export * from './formbuilder/baseform';
 
 export * from './panelview';
 export * from './toolbar';
